리뷰/무비리스트 대폭 수정
리뷰 작성시 한글 깨짐 오류 해결(encoding) (22.07.12)=>
리뷰 중복 작성 불가(22.07.12)
mvc1 패턴 이용해서 메인 페이지에 영화 DB에 따른 모달 출력
해당 영화 이미지 클릭 시 해당 모달 출력하는 것 아직 안돼서 수정쌤께 질문(22.07.12)=>

+++ src/main/java/model/ReviewDeleteContoller.java
@@ -19,8 +19,10 @@ public class ReviewDeleteContoller extends HttpServlet{
 		String idx = req.getParameter("idx");
 			ReviewDAO dao = new ReviewDAO();
 			ReviewDTO dto = dao.selectReview(idx);
-			
-			//삭제-25. 게시물을 삭제
+		HttpSession session = req.getSession();
+		System.out.println(session.getAttribute("UserId").toString());
+		System.out.println(dto.getMembercode());
+		if(session.getAttribute("UserId").toString().equals(dto.getMembercode())) {
 			int result = dao.deleteReview(dto);
 			dao.close();
 			
@@ -30,7 +32,12 @@ public class ReviewDeleteContoller extends HttpServlet{
 			}else {
 				JSFunction.alertBack(resp, "삭제에 실패하였습니다.");
 			}
+		}else {
+			JSFunction.alertBack(resp, "본인이 작성한 리뷰만 삭제 가능합니다!");
+
+			}
 		}
+		
 	}
